Working init for NTVM (actually working one)
